transform: extract the MIR generator and fold oracle into a `mirgen` feature
The random `MirRelationExpr` generators and the `FoldConstants`
result-equivalence oracle lived in the `mz-transform-fuzz` crate, which
sits in the nightly-only cargo-fuzz workspace and so cannot be depended
on from the main workspace. A second consumer now needs them: a suite
that renders the same generated plans on a real replica and checks them
against the same oracle, testing MIR-to-dataflow where the fuzz targets
test MIR-to-MIR.

Move them to `mz_transform::mirgen`, behind a `mirgen` feature that is
off by default. Generation draws from an `Entropy` trait rather than
libFuzzer's `Unstructured` directly, so a byte-driven fuzz target and a
seeded PRNG can both drive it. The fuzz crate keeps its public API and
contributes the `Unstructured` implementation.

`FuzzEntropy` delegates one-to-one, in call order, so byte consumption
is unchanged. That is load-bearing rather than incidental: a corpus
entry is a byte string, which plan it denotes depends on how many bytes
each draw consumes, and release qualification carries a minimized corpus
between runs. A different draw sequence would silently remap every
stored entry and discard the accumulated coverage. Verified by running
both generators over 4000 identical inputs and comparing the produced
plan and the bytes left unconsumed.

Also split the oracle's verdict into `FoldOutcome::{Rows, Error,
Unfoldable}`. `fold_to_multiset` returned `None` both for a plan that
folded to an `EvalError` and for one that did not fold at all, and
callers skipped both. An erroring plan has a definite expected result,
namely the same error, so collapsing the two dropped every
error-propagating plan out of the oracle's reach. This matters for the
generators specifically, since `gen_scalar` emits `DivisionByZero`
poison literals precisely to exercise error propagation.
`fold_to_multiset` keeps its old signature and behaviour.
